% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rstringtie.R
\name{calculateFoldchange}
\alias{calculateFoldchange}
\title{Calculating the fold change}
\usage{
calculateFoldchange(object, genes, crossVar = "condition")
}
\arguments{
\item{object}{An object of test}

\item{genes}{Specify the genes for the calculation of foldchanges}

\item{corssVar}{Default as "condition"}
}
\description{
TE-chimeric exon usage fold changes are calculated based on the coefficients of the GLM fit
}
