% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rstringtie.R
\name{countAnno}
\alias{countAnno}
\title{Counting reads}
\usage{
countAnno(
  annotation,
  bamfile,
  isPairedEnd = TRUE,
  strandSpecific = 0,
  allowMultiOverlap = TRUE,
  ...
)
}
\arguments{
\item{annotation}{Flatted GFF file.}

\item{bamfile}{BAM files path}

\item{isPairedEnd}{A logical scalar or a logical vector, indicating whether 
libraries contain paired-end reads or not.}

\item{strandSpecific}{an integer vector indicating if strand-specific read 
counting should be performed. See also \link[Rsubread]{featureCounts}}

\item{allowMultiOverlap}{logical indicating if a read is allowed to be 
assigned to more than one feature (or meta-feature) if it is found to 
overlap with more than one feature (or meta-feature). TRUE by default.}

\item{...}{Additional arguments. See in \link[Rsubread]{featureCounts}}
}
\description{
Counting the number of reads that overlap with each of the exon
counting bings defined in the flatted GFF file.
}
