% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rstringtie.R
\name{diffExons}
\alias{diffExons}
\title{Differential expression with flatten exons}
\usage{
diffExons(object, design = vector(), ...)
}
\arguments{
\item{object}{RangedSummarizedExperiment object}

\item{design}{Expreiment design}

\item{...}{Ohter arguments}
}
\description{
Diffenential expression with exons
}
