% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rstringtie.R
\name{extractTest}
\alias{extractTest}
\title{Extract the result from the differentially expressed TE-chimeric exon test}
\usage{
extractTest(object)
}
\arguments{
\item{object}{An object of DE TE-chimeric exon test, out from \bold{ChimericDrivenTest}.}
}
\description{
Extract the result from the differentially expressed TE-chimeric exon test
}
