% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rstringtie.R
\name{gffcompareAnno}
\alias{gffcompareAnno}
\title{R wrapper to Run gffcompare tool}
\usage{
gffcompareAnno(reference, gtffile, outfile, params = "")
}
\arguments{
\item{reference}{Use a reference annotation file to guide compare assembly process.}

\item{gtffile}{Gtf files with gffcompare annotation transcripts.}

\item{outfile}{Outfile for the annotation transcripts GTF.}

\item{params}{Other parameter}
}
\description{
The function can be used to compare, merge, annotate and estimate
accuracy of one or more GFF files (the “query” files), when compared with a
reference annotation (also provided as GFF).
}
