% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rstringtie.R
\name{prepareAnno}
\alias{prepareAnno}
\title{Preparing the genome annotation object}
\usage{
prepareAnno(gtffile, singleGens = TRUE, transposone = NULL, minoverlap = 10)
}
\arguments{
\item{gtffile}{GTF file.}

\item{singleGens}{Methods of the disjoint parts, default is TRUE}

\item{minoverlap}{Minimum overlap wit findoverlap, default is 10.}

\item{transposne}{A GRanges object with transposones data}
}
\description{
GTF file, many exons appear multiple times, once for each transcripts
that contains them, need to 'collapse' this information to define exon counting bins.
}
