% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rstringtie.R
\name{stringtieCombine}
\alias{stringtieCombine}
\title{Consolidataion of information on transcript assembly of multiple samples}
\usage{
stringtieCombine(reference = NULL, bamfile = NULL, outfile = NULL, params = "")
}
\arguments{
\item{reference}{Compared with a reference annotation files}

\item{bamfile}{Bamfile, must be of SAM/BAM/CRAM format, sorted by their genomic location.}

\item{outfile}{Output GTF file}

\item{params}{Other parameters}
}
\description{
Transcript quantification is prerequisite for many downstream investigations.
 Several metrics have been proposed for measuring transcript abundance levels based on
 RNA-seq data.
}
