% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rstringtie.R
\name{stringtieMerge}
\alias{stringtieMerge}
\title{R wrapper to Run stringtie tool}
\usage{
stringtieMerge(reference, gtffile, outfile, params = "")
}
\arguments{
\item{reference}{Use a reference annotation file to guide assembly process}

\item{gtffile}{Gtf files with stringtie assemble transcripts [character vector]}

\item{outfile}{Outfile for the merged transcripts GTF}

\item{params}{Other parameter}
}
\description{
Transcript merge mode
}
