% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rstringtie.R
\name{countAnno}
\alias{countAnno}
\title{Counting reads on the exon}
\usage{
countAnno(
  annotation,
  bamfile,
  isPairedEnd = TRUE,
  strandSpecific = 0,
  allowMultiOverlap = TRUE,
  ...
)
}
\arguments{
\item{annotation}{Flattened GFF object.}

\item{bamfile}{Path for BAM files.}

\item{isPairedEnd}{A logical scalar or a logical vector, indicating whether
libraries contain paired-end reads or not.}

\item{strandSpecific}{An integer vector indicating if strand-specific read
counting should be performed. See \code{\link[Rsubread]{featureCounts}}}

\item{allowMultiOverlap}{Logical indicating if a read is allowed to be
assigned to more than one feature (or meta-feature) if it is found to
overlap with more than one feature (or meta-feature). TRUE by default.}

\item{...}{Additional arguments. See \code{\link[Rsubread]{featureCounts}}}
}
\description{
Counting the number of reads that falls into each of the exon
counting bins defined in the flattened GFF object.
}
