% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rstringtie.R
\name{gffcompareAnno}
\alias{gffcompareAnno}
\title{R wrapper to Run gffcompare}
\usage{
gffcompareAnno(reference, gtffile, outfile, params = "")
}
\arguments{
\item{reference}{Use a reference annotation file to guide compare assembly process.}

\item{gtffile}{GTF files with gffcompare annotation transcripts.}

\item{outfile}{The name of the output annotated merged GTF.}

\item{params}{Other parameters}
}
\description{
The function to compare and merge accuracy of one or more GFF files (the “query” files),
when compared with a reference annotation (also provided as GFF).
}
