% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rstringtie.R
\name{stringtieMerge}
\alias{stringtieMerge}
\title{R wrapper to Run stringtie tool}
\usage{
stringtieMerge(reference, gtffile, outfile, params = "")
}
\arguments{
\item{reference}{Use a reference annotation file to guide assembly process}

\item{gtffile}{Character, GTF files assembled from transcripts by stringtie}

\item{outfile}{Character, the output of the merged GTF}

\item{params}{Other parameters}
}
\description{
Merge transcripts
}
